Control if output of a field using ww3_ounf goes to a separate netCDF file
The purpose of this feature is to allow the developer to modify the program ww3_ounf to output tightly connected subfields of a field to the same netCDF file even if the user has set the flag TOGETHER==.FALSE.. The feature itself has no effect on the behavior of the program. The feature introduces a flag FLSEP and and a referencing integer NCJID explained below. The feature also replaces the GOTO loop over partitions with a WHILE () DO .. END DO construction.

FLSEP is introduced as a logical flag indicating that a field should be put in a separate netCDF file. This is one of the two implications of the flag FLFRQ for spectral fields with a third dimension frequency ('f'). FLSEP may be used by the developer to overrule a TRUE value of FLFRQ or .NOT.TOGETHER.

NCJID (ID of a requested 'Joint netCDF' file) may be applied for a field that contains two or more subfields with different units or scales.

To manage such subfields, the loop over partition variables is rewritten to loop over the subfields as well. A loop counter IFP equals IPART+1 when the loop is over partitions. Any such subfield with IFP >= 2, which is NOT a swell partition, is put in the same file as the field with IFP == 1, even if TOGETHER is .FALSE. In the code, NCJID is set in two lines:
  NCJID = 0
  IF ( IFP .GT. 1 .AND. FLSEP .AND. IPART .EQ. 0 ) NCJID = NCIDS(IFI, IFJ, 1)
At this point in the code, the referenced netCDF file NCIDS(IFI, IFJ, 1) is already open for output.

An example of this is applied in the code for the two subfields of the field P2S (IFI,IFJ = 6,7). As these subfields already go to the same netCDF file even when TOGETHER is .FALSE., this change doesn't affect the output except that the standard output messages 'Writing new record ...' also reports subfields as separate records.

As a side example FLSEP could apply to the field USP (IFI,IFJ = 6,12) as follows in two settings a, b. This  wouldn't change the actual output except for the dimension name.

a) Specify a name of the third dimension, e.g. segmented wave frequencies DIM3NAME(1:2)='fs' (for how to further modify the code, look for 'fs' in the comments lines).

b) Set FLSEP=.TRUE. to indicate that the field must be in a separate netCDF file (similar to FLFRQ=.TRUE. when DIM3NAME=='f'). If you choose FLSEP=.FALSE., the field can go to a common netCDF if mode TOGETHER is chosen by the user.