Allow a 'spectral' third dimension other than 'f' in netCDF output
The general aim of this change to the program ww3_ounf is to allow a field with a new 'third' dimension to be written to netCDF together with other fields. The present feature introduces no change to the users input or the output. The main effect is that an extra stage NF90_REDEF(NCID) ... NF90_ENDDEF(NCID) may occur during the processing.

An extra 'third' dimension may be named e.g. segmented frequencies 'fs'. One third dimension, frequency 'f', already exists in the code, and a field with an 'f' dimension is forced to go to a separate netCDF. With this feature it is still so that if developer defines two different third dimensions that share their netCDF name, the fields that depend on the dimension are forced to go to separate files.

The present feature depends on the feature fb_ounf_jid ("Control if output of a field using ww3_ounf goes to a separate netCDF").

The developer must use a new subroutine W3NCDEF3 to define the netCDF attributes of third dimension variables like 'f' or 'fs'.

 To manage the possibility of more than one third dimension, and possibly in a file together with other output fields, the indexing of of the program's storage arrays DIMID, DIMFIELD and VARID is rewritten to avoid extensive algebraic expressions directly when addressing each dimension. For this we introduce two integers ITD: the index in DIMID of the time dimension, and TDIM: the index in DIMFIELD of the time dimension.

The values of ITD and TDIM can be determined from the following table:

COORD TYPE:      Dims in DIMID:                             Dims in DIMFIELD:
Regular          (level,    x,    y,   [f,]  time)          (x, y, [f,] time)
SMC flat         (level, cell,   (),   [f,]  time)          (cell, [f,] time)
Unstructured     (level, node, elem,   [f,]  time,  noel)   (node, [f,] time)
INDEX in DIMID:            2     3   [ITD-1]  ITD   ITD+1
INDEX in VARID:            1     2      []   ITD-1   ITD
(The 'level' is not a dimension of any actual field and not in DIMFIELD)

By default the field has no third dimension and ITD = 4. If it has a third dimension (e.g. frequency 'f' in FIELD(x,y,f,time)), then ITD is shifted by one to ITD = 5. DIMID(ITD-1) is the netCDF ID of the third dimension.

If there is no third dimension, the time index in DIMFIELD is TDIM = 3 for regular grid, or TDIM = 2 for an unstructured or a flat SMC grid.
TDIM is shifted by one if the variable has a third dimension

The present feature further simplifies the code for defining and storing the fields into the netCDF files. There are two simplifications, 1 and 2:

Simplification 1) to sections 2.6.1 "Get the dimensions from the netcdf header" and 2.6.3 "Defines or gets the variables identifiers".

This part of the code is reduced from about 270 lines to about 180 lines.

Using the the new index TDIM we avoid conditional blocks depending on the grid being SMC flat, or unstructured. The only condition that remains is if the variable type is REAL or a SHORT integer, so that e.g. the definition of a field in the natCDF file reads:

        IF (NCVARTYPE.EQ.2) THEN
          IRET=NF90_DEF_VAR(NCID,VARNM(I), NF90_SHORT, DIMFIELD(1:TDIM), VARID(IVAR))
        ELSE
          IRET=NF90_DEF_VAR(NCID,VARNM(I), NF90_FLOAT, DIMFIELD(1:TDIM), VARID(IVAR))
        END IF
!/NC4        IF (NCTYPE.EQ.4) IRET = NF90_DEF_VAR_DEFLATE(NCID, VARID(IVAR), 1, 1, DEFLATE)
             CALL CHECK_ERR(IRET)

The code is restructured so that the fields definition above only occurs in one block in the code: at the first time step. This block is delayed to a stage after writing the values of dimension variables and MAPSTAT and thus involves an extra stage of netCDF define mode (starting with function NF90_REDEF(NCID) and ending with NF90_ENDDEF(NCID)).

Simplification 2) to the section 2.6.5 "Puts field(s) in NetCDF file"

This part of the code is reduced from about 500 lines to about 140 lines.

There are six different types of fields to put into the netCDF depending on the var type being REAL or SHORT and the number of subfields being one, two or three. For each of these types there is a call of the funcion NF90_PUT_VAR(). This is a new simplification of the code, because there is further four different combinations of field types depending on there being a third dimension ('f') or not (EXTRADIM==1 or 0) in combination with the grid being an SMC flat grid or not (SMCTYPE==1 or 2). These four types are now combined before calling the function NF90_PUT_VAR(). This is done by copying the subfield into a common array (e.g. X1 when REAL type, MX1 for SHORT integer), and by determining the function arguments START(1:TDIM) and COUNT(1:TDIM) of NF90_PUT_VAR() from the following principle, where NTIM is the current time step and TDIM is the index in DIMID as explained above:

      ! For a regular grid ( SMCTYPE == 2 or .NOT.UNGTYPE ) we have
      ! if EXTRADIM==0: TDIM=3; START=(/1,1,NTIM/); COUNT=(/NX,NY,1/)
      ! if EXTRADIM==1: TDIM=4; START=(/1,1,IK,NTIM/); COUNT=(/NX,NY,1,1/)
      !
      ! For flat sea point array, let NX == COUNT(1) = IXN-IX1+1. We have:
      ! if EXTRADIM==0: TDIM=2; START=(/1,NTIM/); COUNT=(/NX,1/)
      ! if EXTRADIM==1: TDIM=3; START=(/1,IK,NTIM/); COUNT=(/NX,1,1/)

If there is a third dimension we loop over elements IK=I1F,I2F. If there is no third dimension this loop reduces to one step IK=0. Within the loop the field is put into the netCDF by the statement

      IRET=NF90_PUT_VAR(NCID,VARID(IVAR1), &
              X1(IX1:IXN,IY1:IYN),START(1:TDIM),COUNT(1:TDIM))