In ww3_ounf, let sub-parameters go to a joined netCDF even if TOGETHER is false
The purpose of this feature is to facilitate output output of tightly connected sub-parameters to the same netCDF file even if the user has set the flag TOGETHER=.FALSE.. The feature itself has no effect on the behavior of the program. The feature introduces a flag FLJOIN explained below. The feature also replaces the GOTO loop over partitions with a WHILE () DO .. END DO construction.

After applying this feature to two specific parameters (BED and P2S) the if-then-else construct looped within IFI,IFJ (which constitutes a hard-coded configuration) now results strictly in one of three types of horizontal fields: A scalar (e.g. hs) with NFIELD==1 in array X1(1:NX+1,1:NY), a horizontal vector (e.g. wnd) with NFIELD==2 in arrays(XX,XY), or a tensor (sxy) with NFIELD==3 in arrays (X1,X2,XY). The same types apply if a third dimension (e.g.frequency) is present, with scalars to XK(1:NX+1,1:NY,1:NK) and vectors to (XXK, XYK).

A logical flag FLJOIN is introduced to loop over a number of such sub-parameters. FLJOIN indicates if a sub-parameter should be put together in a joint netCDF file.

Currently the maximum number of sub-parameters in the code is NSPMAX=2.

A field parameter may contain more than one sub-parameters. Two examples of this are the parameter 'P2S' (IFI,IFJ=6,7), which contains two scalar sub-parameters 'p2s' and 'pp2s', and the parameter 'BED' (IFI,IFJ=7,3), which  contains two sub-parameters, a scalar 'bed' and a vector ('ripplex', 'rippley').

The flag FLJOIN may be set to True for individual sub-parameters in the hardcoded configuration, and in case the overall input (namelist) parameter NML_FIELD%SAMEFILE is False the sub-parameter will be put together with the first sub-parameter under the common parameter.

To manage this, a counter ISP=1,NSP is applied in a loop over these sub-parameters. In addition ISP also acts as a counter over swell partitions. Any sub-parameter with ISP >= 2, which has a configured True value of FLJOIN, is put in the same file as the parameter with ISP == 1, even if TOGETHER is .FALSE. A message 'Writing the new record ...' is written to standard output for each sub-parameter.

Note that FLJOIN is NOT set to True for swell partitions. FLJOIN is always set to False for the parameters that are spectral fields over three dimensions, x, y, and frequency ('f'), for which is a configuration flag FLFRQ is True.

A parameter with ISP == 1 can be forced to a separate file by setting FLJOIN = .FALSE., but forcing FLJOIN to True only makes sense for ISP > 1.