Sub-fields placed in a joined file even if TOGETHER is False
In the output to netCDF files, a field may contain more than one sub-fields.
Two examples of this are the fields 'P2S' (IFI,IFJ=6,7) and 'BED' (IFI,IFJ=7,3).
'P2S' contains two scalar sub-fields 'p2s' and 'pp2s'. 'BED' contains
two sub-fields, a scalar 'bed' and a 2D field ('ripplex', 'rippley').

The maximum number of sub-fields is NFSMAX=2 as presently hard-coded
in the SUBROUTINE INIT_META of w3ouncmetamd.F90.

FLJOIN is a flag indicating if a sub-field should be put together in
a joint netCDF file, even if TOGETHER is .FALSE.. It is set to False in
the case of a spectral field over four dimensions (time, f, y, x) with
f the frequency ('f') for which a configuration flag FLFRQ is True. For a
lat-lon grid y is 'latitude' and x is 'longitude'.

The flag FLJOIN may be set to True for individual sub-fields in the
hardcoded configuration, and in case the overall input (namelist) parameter
NML_FIELD%SAMEFILE is False, the sub-field will be put together with
the first sub-field under the common field index pair (IFI,IFJ).

To manage this, a counter IFS=1,NFS is applied in a loop over these
sub-fields and a sub-field with IFS >= 2 may have a configured True
value of FLJOIN. Alternatively IFS acts as a counter over swell
partitions. However, FLJOIN is NOT set to True for swell partitions.
Messages 'Writing the new record <(sub-)field> number ...' are
written to standard output.

Note: A field with IFS == 1 can be forced to a separate file by setting
FLJOIN to False. You cannot do the reverse: If you try to set FLJOIN to
True for IFS == 1, it will be changed to the logical value of TOGETHER.