More than one different 'third' dimensions in the same netCDF file
A new 'third' dimension (like 'f' for frequency after the 2D
spatial dimensions) may be added to the code, for example
segmented frequencies DIM3NAME='fs'. In that case you may also
set FLJOIN=.TRUE. (thereby outputting the variable in a file
TOGETHER with other variables). Ensure that FLFRQ=.FALSE. . Note
however, it is bad practice to have two different 'frequency'-like
dimensions in the same netCDF file.

The netCDF attributes for a third dimension variable (like 'f')
are defined in the netCDF file using the subroutine W3NCDEF3.
The procedure is constructed so that when such a third dimension
is requested and FLJOIN is not False, the variable is put in a
netCDF file together with other variables. However, if a another
third dimension with the same name is requested for another
variable, output is forced to a separate netCDF file.