fb_xstp: New &XSTP namelist statement to define the spectral tail extension
Purpose: For most output fields the tail of the variance spectrum added beyond the highest prognostic frequency SIG(NK) is by default assumed to decay with frequency as SIG^{-4}, and have a uniform directional distribution over all frequencies. For some fields, however, an integration over this tail does not converge as SIG goes to infinity. For example, the value of the Stokes drift at the very surface, uss, is calculated as an integral over the spectrum times SIG^3. Historically, other parameters such as the 'surface roughness' as seen on radar backscatter, or the Goda peakedness parameter in estimation of the Benjamin-Feir instability, would involve an integral over the spectrum times SIG^4 if directionality is not taken into account.

Therefore, when we wish to calculate a parameter that is very sensitive to the tail form, we need a refinement of the spectral shape. That may be a tail with a transition to a SIG^{-5} decay, or assuming the directional spread increases with frequency, or mimicking a high frequency cut-off in the observation technique.

The &XSTP namelist group is specified in the input to ww3_grid. To output fields assuming a specific spectral tail shape, you run ww3_grid to create a specific mod_defX.ww3 file.

Files changed:

model/src/w3gdatmd.F90: Declare and initialize the model parameters
model/src/w3gridmd.F90: Declare the namelist and translate to model parameters
                        e.g. NAMELIST /XSTP/ USXTT, USXFM
model/src/w3iogrmd.F90: Write/read the model parameters to mod_defX.ww3
model/src/w3iogomd.F90: Calculation during model run

! Implemented namelist parameters and defaults:
NAMELIST /XSTP/ USXTT, USXFM
USXTT='none' ! Tail type (C*4)
! Implemented tail types:
! USXTT='none': Cut-off at NK.
! USXTT='Pf-5': Spectrum of Stokes drift with a SIG^{-2} tail as for a
!               variance spectrum with a Phillips (1958) freq^{-5} tail
!               and a constant directional spread.
USXFM=10     ! Cut-off frequency (Real), units: Hz.
             ! USXF >= 10 Hz means integration to infinity.