Bugfix for calculation of Goda's (1970) peakedness parameter QP
This parameter is still investigated, mainly for verification of
modelling or observation systems. See e.g. Le Merle et al. (2021),
Magnusson et al. (2021), or Ewans et al. (2021).

The bugfic corrects errors in the formulations for the field QP as
well as for the partitioned fields PQP.

Steps to reproduce the error:

1) Choose QP and PQP for output fields and run a realistic case:
Although the bulk of output values of QP are in the expected range
of 1 to 5 (dimensionless), the code yields many extreme outliers.

2) Change the the number of direction bins in the model definition
mod_def.ww3, e.g. from 36 to 24 or vice versa: The output fields of
QP shift approximately proportionally to the directional bin width DTH

There are corrections in w3iogomd.F90 and w3partmd.F90.

The corrections are explained in the code in comments starting with
"! C Hansen: ". These comments may be modified when the bugfix is
accepted or corrected.